+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.115.1 - 2024-03-15
+
+- chore: remove v1 rrweb loading (#1078)
+
+## 1.115.0 - 2024-03-14
+
+- feat: track recording URL without pageview capture (#1076)
+- fix: return typing of global functions (#1081)
+
 ## 1.114.2 - 2024-03-12
 
 - fix: patch rrweb zero width canvas bug (#1075)